Add files via upload
Had to make a couple vital changes to the script. After Situation 3, it said that if the player chose 1.3 & 3.1, he'd get his toes ripped off. I changed this to 1.1, which would make the most sense cuz flip flops. Also added END at the end of the HTML bit, to signify that it's a bad end.