A LOT (pls see description)
Added an image. Unfortunately can't get it to cycle through at the moment. Will keep this for now as having an image was not a REAL requirement. It would be nice if we can figure it out though.

A LOT of formatting has been put into the buttons and the panel. Made the Right panel smaller and set a fixed size on Right panel's buttons.

I formatted ALL text on the buttons to look nicer.